move ui components to ui folder
