+++ src/components/pages/members.tsx
@@ -1,6 +1,6 @@
 import { Card, CardContent } from "@/components/ui/card";
 import { PageLayout } from "../layouts/page-layout";
-import { DataTable } from "../members-table";
+import { DataTable } from "../ui/members-table";
 
 export function Members() {
   return (
