+++ src/components/ui/account-details.tsx
@@ -11,12 +11,12 @@ import {
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 import { Globe, LogOut, Moon, Sun } from "lucide-react";
 import { useAuth } from "@/hooks/useAuth";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useTheme } from "./theme-provider";
+import { useTheme } from "../theme-provider";
 
 export function AccountDetails() {
   const { user, signOut } = useAuth();
