+++ src/components/ui/login.tsx
@@ -7,13 +7,13 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from "./ui/form";
+} from "./form";
 
-import { Input } from "./ui/input";
+import { Input } from "./input";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 import { useAuth } from "@/hooks/useAuth";
 import { useTranslation } from "react-i18next";
