+++ src/components/ui/reset-button.tsx
@@ -9,7 +9,7 @@ import {
   AlertDialogTitle,
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
-import { Button } from "./ui/button";
+import { Button } from "./button";
 
 export function ResetButton({
   resetForm,
